refactor soup service sharing to use Arc directly
