+++ crates/complete_graph/src/lib.rs
@@ -18,7 +18,6 @@ pub use graphql_properties::{
     PropertiesEntityPropertyWriter, PropertiesMutationRoot, PropertiesSoupPropertyEdgeReader,
     SoupPropertyEdgeReader, entity_properties_loader,
 };
-pub use graphql_soup::SharedSoupService;
 pub use schema::{
     SchemaOnlySoupSchema, SchemaOnlyState, SharedSoupSchema, SoupQueryRoot, SoupSchema,
     build_schema, build_schema_from_arc, build_schema_with_service,
